fix(CamelYamlDsl): Fix Camel Yaml DSL union types
In Camel 4.15, the `default` values are written in `string`, regardless
of its type. For instance, this is how the `scriptPropertyPlaceholders`
property is defined, notice the `default: "true"` portion.

```
          "scriptPropertyPlaceholders" : {
            "type" : "boolean",
            "title" : "Script Property Placeholders",
            "description" : "Whether the script should support using Camel property placeholder syntax {{ }}.",
            "default" : "true"
          },
```

This is problematic because it generates the followng `typescript`
definition:
```
scriptPropertyPlaceholders: boolean | string
```

Thix will be fixed in Camel upstream, but in the meantime, we need to
provide a workaround for the already published versions.